+++ packages/@leight/user/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./api";
-export * from "./context";
 export * from "./error";
 export * from "./schema";
 export * from "./source";