+++ packages/@leight/container/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/container
 
+## 0.5.31
+
+### Patch Changes
+
+- Improved user/container stuff
+
 ## 0.5.30
 
 ### Patch Changes