+++ packages/@leight/next.js-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/next.js-server
 
+## 0.5.36
+
+### Patch Changes
+
+- Improved user/container stuff
+- Updated dependencies
+    - @leight/user-server@0.5.34
+    - @leight/container@0.5.31
+    - @leight/user@0.5.34
+
 ## 0.5.35
 
 ### Patch Changes