+++ packages/@leight/trpc-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/trpc-server
 
+## 0.5.33
+
+### Patch Changes
+
+- Improved user/container stuff
+- Updated dependencies
+    - @leight/user-server@0.5.34
+    - @leight/container@0.5.31
+    - @leight/user@0.5.34
+
 ## 0.5.32
 
 ### Patch Changes