+++ packages/@leight/user-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/user-server
 
+## 0.5.34
+
+### Patch Changes
+
+- Improved user/container stuff
+- Updated dependencies
+    - @leight/container@0.5.31
+    - @leight/user@0.5.34
+
 ## 0.5.33
 
 ### Patch Changes