+++ packages/@leight/user/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/user
 
+## 0.5.34
+
+### Patch Changes
+
+- Improved user/container stuff
+- Updated dependencies
+    - @leight/container@0.5.31
+
 ## 0.5.33
 
 ### Patch Changes